SONARJAVA-4252 SONARJAVA-4336 Move Java 9 code samples to compiled folder
