+++ java-checks/src/test/java/org/sonar/java/checks/MutableMembersUsageCheckTest.java
@@ -22,25 +22,25 @@ package org.sonar.java.checks;
 import org.junit.jupiter.api.Test;
 import org.sonar.java.checks.verifier.CheckVerifier;
 
+import static org.sonar.java.checks.verifier.TestUtils.mainCodeSourcesPath;
 import static org.sonar.java.checks.verifier.TestUtils.nonCompilingTestSourcesPath;
-import static org.sonar.java.checks.verifier.TestUtils.testSourcesPath;
 
 class MutableMembersUsageCheckTest {
 
   @Test
   void test() {
     CheckVerifier.newVerifier()
-      .onFile(testSourcesPath("checks/MutableMembersUsageCheck.java"))
+      .onFile(mainCodeSourcesPath("checks/MutableMembersUsageCheck.java"))
       .withCheck(new MutableMembersUsageCheck())
       .verifyIssues();
   }
 
   @Test
-  void test_java_9() {
+  void test_non_compiling() {
     CheckVerifier.newVerifier()
       .onFile(nonCompilingTestSourcesPath("checks/MutableMembersUsageCheck.java"))
       .withCheck(new MutableMembersUsageCheck())
-      .verifyIssues();
+      .verifyNoIssues();
   }
 
 }
